Share OtlpEndpointResolver
